completed with limited styling

+++ client/src/components/Note.vue
@@ -1,13 +1,11 @@
 <template>
-  <div class="note">
-    <tr>
-      <td>{{notesData.reportedBy}}</td>
-      <td>{{notesData.content}}</td>
-      <td>
-        <i @click="deleteNote" class="fas fa-trash"></i>
-      </td>
-    </tr>
-  </div>
+  <tr class="note">
+    <td>{{notesData.reportedBy}}</td>
+    <td>{{notesData.content}}</td>
+    <td>
+      <i @click="deleteNote" class="fas fa-trash"></i>
+    </td>
+  </tr>
 </template>
 
 <script>
@@ -16,7 +14,20 @@ export default {
   props: ["notesData"],
   methods: {
     deleteNote() {
-      this.$store.dispatch("deleteNote", this.notesData.id);
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then(result => {
+        if (result.value) {
+          this.$store.dispatch("deleteNote", this.notesData.id);
+          Swal.fire("Deleted!", "Your file has been deleted.", "success");
+        }
+      });
     }
   }
 };
